+++ test/unit/android/device/mjsonwp/event_test.rb
@@ -15,12 +15,12 @@
 require 'test_helper'
 require 'webmock/minitest'
 
-# $ rake test:unit TEST=test/unit/android/device/mjsonwp/logs_test.rb
+# $ rake test:unit TEST=test/unit/android/device/mjsonwp/event_test.rb
 class AppiumLibCoreTest
   module Android
     module Device
       module MJSONWP
-        class LogsTest < Minitest::Test
+        class EventTest < Minitest::Test
           include AppiumLibCoreTest::Mock
 
           def setup
@@ -34,6 +34,16 @@ class AppiumLibCoreTest
               .to_return(headers: HEADER, status: 200, body: { value: nil }.to_json)
 
             @driver.logs.event vendor: 'appium', event: 'customEvent'
+
+            assert_requested(:post, "#{SESSION}/appium/log_event", times: 1)
+          end
+
+          def test_log_event_equal
+            stub_request(:post, "#{SESSION}/appium/log_event")
+              .with(body: { vendor: 'appium', event: 'customEvent' }.to_json)
+              .to_return(headers: HEADER, status: 200, body: { value: nil }.to_json)
+
+            @driver.logs.event = { vendor: 'appium', event: 'customEvent' }
 
             assert_requested(:post, "#{SESSION}/appium/log_event", times: 1)
           end
@@ -57,7 +67,7 @@ class AppiumLibCoreTest
 
             assert_requested(:post, "#{SESSION}/appium/events", times: 1)
           end
-        end # class Logs
+        end # class EventTest
       end # module MJSONWP
     end # module Device
   end # module Android