+++ CHANGELOG.md
@@ -5,9 +5,8 @@ Read `release_notes.md` for commit level details.
 ## [Unreleased]
 
 ### Enhancements
-- Add `Logs#event` to post a custom log. `@driver.logs.event vendor: 'appium', event: 'funEvent'`
-- Add `Logs#events` to get events. `@driver.logs.events` is the command.
-    - It equals to  `@driver.session_capabilities['events']`
+- Add `Logs#event` to post a custom log by `@driver.logs.event vendor: 'appium', event: 'funEvent'`
+- Add `Logs#events` to get events by `@driver.logs.events`. It is equal to  `@driver.session_capabilities['events']`
 
 ### Bug fixes